--- FrontEnd/src/styles/pages/user/_user_header.scss
@@ -1,3 +1,0 @@
-.logined-user {
-  display: flex;
-}
